refactor(protocol): split gateway protocol registries
