+++ packages/gateway-protocol/src/type-export-registry.ts
@@ -1,0 +1,544 @@
+// Type exports mirror the schema exports for downstream TypeScript consumers.
+export type {
+  GatewayFrame,
+  ConnectParams,
+  WorkerAdmissionFailureReason,
+  WorkerAdmissionHandshake,
+  WorkerAdmissionResponseFrame,
+  WorkerConnectParams,
+  WorkerConnectRequestFrame,
+  WorkerErrorShape,
+  WorkerHeartbeatParams,
+  WorkerHeartbeatRequestFrame,
+  WorkerHeartbeatResult,
+  WorkerHeartbeatResponseFrame,
+  WorkerHelloOk,
+  WorkerLiveEvent,
+  WorkerLiveEventErrorDetails,
+  WorkerLiveEventErrorShape,
+  WorkerLiveEventParams,
+  WorkerLiveEventRequestFrame,
+  WorkerLiveEventResponseFrame,
+  WorkerLiveEventResult,
+  WorkerProtocolCloseReason,
+  WorkerTranscriptCommitErrorReason,
+  WorkerTranscriptCommitErrorShape,
+  WorkerTranscriptCommitParams,
+  WorkerTranscriptCommitRequestFrame,
+  WorkerTranscriptCommitResponseFrame,
+  WorkerTranscriptCommitResult,
+  WorkerTranscriptMessage,
+  GatewaySuspendTaskBlocker,
+  GatewaySuspendBlocker,
+  GatewaySuspendPrepareParams,
+  GatewaySuspendPrepareResult,
+  GatewaySuspendStatusParams,
+  GatewaySuspendStatusResult,
+  GatewaySuspendResumeParams,
+  GatewaySuspendResumeResult,
+  HelloOk,
+  RequestFrame,
+  ResponseFrame,
+  EventFrame,
+  PresenceEntry,
+  Snapshot,
+  ErrorShape,
+  StateVersion,
+  AgentEvent,
+  ConversationListItem,
+  ConversationListParams,
+  ConversationListResult,
+  ConversationSendParams,
+  ConversationSendResult,
+  ConversationTurnCancelParams,
+  ConversationTurnCancelResult,
+  ConversationTurnParams,
+  ConversationTurnReply,
+  ConversationTurnResult,
+  AgentIdentityParams,
+  AgentIdentityResult,
+  AgentWaitParams,
+  ChatEvent,
+  ChatRunStartupPhase,
+  ChatStatusEvent,
+  TickEvent,
+  ShutdownEvent,
+  WakeParams,
+  NodePairListParams,
+  NodePairApproveParams,
+  DevicePairListParams,
+  DevicePairApproveParams,
+  DevicePairRejectParams,
+  DevicePairSetupCodeParams,
+  DevicePairSetupCodeResult,
+  DevicePairRenameParams,
+  ConfigGetParams,
+  ConfigSetParams,
+  ConfigApplyParams,
+  ConfigPatchParams,
+  ConfigSchemaParams,
+  ConfigSchemaResponse,
+  SystemAgentChatParams,
+  SystemAgentChatQuestion,
+  SystemAgentChatResult,
+  SystemAgentChatHistoryParams,
+  SystemAgentChatHistoryResult,
+  SystemAgentChatHistoryTurn,
+  SystemChangeEntry,
+  SystemChangeKind,
+  SystemChangeSource,
+  SystemChangesListParams,
+  SystemChangesListResult,
+  SystemAgentSetupDetectParams,
+  SystemAgentSetupDetectResult,
+  SystemAgentSetupVerifyParams,
+  SystemAgentSetupVerifyResult,
+  SystemAgentSetupActivateParams,
+  SystemAgentSetupActivateResult,
+  SystemAgentSetupAuthStartParams,
+  SystemAgentSetupAuthStartResult,
+  WizardStartParams,
+  WizardNextParams,
+  WizardCancelParams,
+  WizardStatusParams,
+  WizardStep,
+  WizardNextResult,
+  WizardStartResult,
+  WizardStatusResult,
+  TalkCatalogParams,
+  TalkCatalogResult,
+  TalkClientCreateParams,
+  TalkClientCreateResult,
+  TalkClientCloseParams,
+  TalkClientMutationResult,
+  TalkClientSteerParams,
+  TalkAgentControlResult,
+  TalkClientToolCallParams,
+  TalkClientToolCallResult,
+  TalkClientTranscriptParams,
+  TalkConfigParams,
+  TalkConfigResult,
+  TalkSessionAppendAudioParams,
+  TalkSessionAcknowledgeMarkParams,
+  TalkSessionCancelOutputParams,
+  TalkSessionCancelTurnParams,
+  TalkSessionCreateParams,
+  TalkSessionCreateResult,
+  TalkSessionJoinParams,
+  TalkSessionJoinResult,
+  TalkSessionTurnParams,
+  TalkSessionTurnResult,
+  TalkSessionSteerParams,
+  TalkSessionSubmitToolResultParams,
+  TalkSessionCloseParams,
+  TalkSessionOkResult,
+  TalkSpeakParams,
+  TalkSpeakResult,
+  TtsSpeakParams,
+  TtsSpeakResult,
+  TalkModeParams,
+  ChannelsStatusParams,
+  ChannelsStatusResult,
+  ChannelsPairingListParams,
+  ChannelsPairingListResult,
+  ChannelsPairingApproveParams,
+  ChannelsPairingApproveResult,
+  ChannelsPairingDismissParams,
+  ChannelsPairingDismissResult,
+  ChannelsPairingAccount,
+  ChannelsPairingRequest,
+  ChannelsStartParams,
+  ChannelsStopParams,
+  ChannelsLogoutParams,
+  WebLoginStartParams,
+  WebLoginWaitParams,
+  AgentKind,
+  AgentSummary,
+  AgentsFileEntry,
+  AgentsCreateParams,
+  AgentsCreateResult,
+  AgentsUpdateParams,
+  AgentsUpdateResult,
+  AgentsDeleteParams,
+  AgentsDeleteResult,
+  AgentsFilesListParams,
+  AgentsFilesListResult,
+  AgentsFilesGetParams,
+  AgentsFilesGetResult,
+  AgentsFilesSetParams,
+  AgentsFilesSetResult,
+  AgentsWorkspaceEntry,
+  AgentsWorkspaceFile,
+  AgentsWorkspaceListParams,
+  AgentsWorkspaceListResult,
+  AgentsWorkspaceGetParams,
+  AgentsWorkspaceGetResult,
+  SessionFileBrowserEntry,
+  SessionFileBrowserResult,
+  SessionFileEntry,
+  SessionFileKind,
+  SessionFileRelevance,
+  SessionsFilesListParams,
+  SessionsFilesListResult,
+  SessionsFilesGetParams,
+  SessionsFilesGetResult,
+  SessionsFilesSetParams,
+  SessionsFilesSetResult,
+  SessionsFilesRevealParams,
+  SessionsFilesRevealResult,
+  SessionDiffFile,
+  SessionDiffFileStatus,
+  SessionsDiffParams,
+  SessionsDiffResult,
+  SessionDiscussionState,
+  SessionDiscussionInfo,
+  SessionDiscussionInfoParams,
+  SessionDiscussionInfoResult,
+  SessionDiscussionOpenParams,
+  SessionDiscussionOpenResult,
+  ArtifactSummary,
+  ArtifactsListParams,
+  ArtifactsListResult,
+  ArtifactsGetParams,
+  ArtifactsGetResult,
+  ArtifactsDownloadParams,
+  ArtifactsDownloadResult,
+  AgentsListParams,
+  AgentsListResult,
+  ChatMetadataParams,
+  ChatToolTitlesParams,
+  ChatToolTitlesResult,
+  CommandsListParams,
+  CommandsListResult,
+  CommandEntry,
+  PluginCatalogEntry,
+  PluginsInstallParams,
+  PluginsInstallResult,
+  PluginsListParams,
+  PluginsListResult,
+  PluginsRefreshParams,
+  PluginsRefreshResult,
+  PluginsSearchParams,
+  PluginsSearchResult,
+  PluginsSessionActionParams,
+  PluginsSessionActionResult,
+  PluginsSetEnabledParams,
+  PluginsSetEnabledResult,
+  PluginsUninstallParams,
+  PluginsUninstallResult,
+  AuthProbeStatus,
+  ModelsProbeParams,
+  ModelsProbeResult,
+  ModelsProbeTargetResult,
+  SkillsStatusParams,
+  ToolsCatalogParams,
+  ToolsCatalogResult,
+  ToolsEffectiveParams,
+  ToolsEffectiveResult,
+  ToolsInvokeParams,
+  ToolsInvokeResult,
+  SkillsBinsParams,
+  SkillsBinsResult,
+  SkillsCuratorActionParams,
+  SkillsCuratorActionResult,
+  SkillsCuratorStatusParams,
+  SkillsCuratorStatusResult,
+  SkillsSearchParams,
+  SkillsSearchResult,
+  SkillsDetailParams,
+  SkillsDetailResult,
+  SkillsProposalsListParams,
+  SkillsProposalsListResult,
+  SkillsProposalInspectParams,
+  SkillsProposalInspectResult,
+  SkillsProposalCreateParams,
+  SkillsProposalUpdateParams,
+  SkillsProposalReviseParams,
+  SkillsProposalRequestRevisionParams,
+  SkillsProposalRequestRevisionResult,
+  SkillsProposalActionParams,
+  SkillsProposalApplyResult,
+  SkillsProposalRecordResult,
+  SkillsSecurityVerdictsParams,
+  SkillsSecurityVerdictsResult,
+  SkillsSkillCardParams,
+  SkillsSkillCardResult,
+  SkillsUploadBeginParams,
+  SkillsUploadChunkParams,
+  SkillsUploadCommitParams,
+  SkillsInstallParams,
+  SkillsUpdateParams,
+  EnvironmentStatus,
+  WorkerEnvironmentState,
+  WorkerTunnelStatus,
+  WorkerEnvironmentMetadata,
+  EnvironmentSummary,
+  EnvironmentsCreateParams,
+  EnvironmentsCreateResult,
+  EnvironmentsDestroyParams,
+  EnvironmentsDestroyResult,
+  EnvironmentsListParams,
+  EnvironmentsListResult,
+  EnvironmentsStatusParams,
+  EnvironmentsStatusResult,
+  SystemInfoParams,
+  SystemInfoResult,
+  NodePairRejectParams,
+  NodePairRemoveParams,
+  NodeListParams,
+  NodePluginToolDescriptor,
+  NodePluginToolsUpdateParams,
+  NodeSkillDescriptor,
+  NodeSkillsUpdateParams,
+  NodeInvokeParams,
+  NodeInvokeInputEvent,
+  NodeInvokeProgressParams,
+  NodeInvokeResultParams,
+  NodeEventParams,
+  NodeEventResult,
+  NodePresenceAlivePayload,
+  NodePresenceAliveReason,
+  NodePresenceActivityPayload,
+  NodePendingDrainParams,
+  NodePendingDrainResult,
+  NodePendingEnqueueParams,
+  NodePendingEnqueueResult,
+  SessionsListParams,
+  SessionsSearchHit,
+  SessionsSearchParams,
+  SessionsSearchResult,
+  SessionsCleanupParams,
+  SessionsPreviewParams,
+  SessionsDescribeParams,
+  SessionsResolveParams,
+  SessionOperationEvent,
+  SessionCompanionExchange,
+  SessionObserverDigest,
+  SessionObserverHealth,
+  SessionObserverPlanProgress,
+  SessionMember,
+  SessionMemberAddParams,
+  SessionMemberMutationResult,
+  SessionMemberRemoveParams,
+  SessionMembersListParams,
+  SessionMembersListResult,
+  SessionSharingAction,
+  SessionSharingEvent,
+  SessionSharingIdentity,
+  SessionSharingRole,
+  SessionVisibility,
+  SessionVisibilitySetParams,
+  SessionVisibilitySetResult,
+  SessionsCompanionAskParams,
+  SessionsCompanionAskResult,
+  SessionsCompanionResetParams,
+  SessionsCompanionResetResult,
+  SessionsCompanionStateParams,
+  SessionsCompanionStateResult,
+  SessionsObserverVisibilityParams,
+  SessionsObserverVisibilityResult,
+  SessionPlacementState,
+  SessionPlacement,
+  SessionWorktreeInfo,
+  SessionsDispatchParams,
+  SessionsDispatchResult,
+  SessionsReclaimParams,
+  SessionsReclaimResult,
+  SessionsCreateResult,
+  SessionBranch,
+  SessionsBranchesListParams,
+  SessionsBranchesListResult,
+  SessionsBranchesSwitchParams,
+  SessionsBranchesSwitchResult,
+  SessionsForkParams,
+  SessionsForkResult,
+  SessionsRewindParams,
+  SessionsRewindResult,
+  SessionsPatchParams,
+  SessionsResetParams,
+  SessionsDeleteParams,
+  SessionsCompactParams,
+  SessionsUsageParams,
+  AuditActivityAgentRunV1,
+  AuditActivityEventV1,
+  AuditActivityInboundMessageV1,
+  AuditActivityListParams,
+  AuditActivityListResult,
+  AuditActivityOutboundMessageV1,
+  AuditActivityToolActionV1,
+  AuditEvent,
+  AuditListParams,
+  AuditListResult,
+  UserProfile,
+  UsersLinkEmailParams,
+  UsersLinkEmailResult,
+  UsersListParams,
+  UsersListResult,
+  UsersSelfParams,
+  UsersSelfResult,
+  UsersSetAvatarParams,
+  UsersSetAvatarResult,
+  UsersSetDisplayNameParams,
+  UsersSetDisplayNameResult,
+  TaskSuggestion,
+  TaskSuggestionEvent,
+  TaskSuggestionResolution,
+  TaskSuggestionsAcceptParams,
+  TaskSuggestionsAcceptResult,
+  TaskSuggestionsCreateParams,
+  TaskSuggestionsCreateResult,
+  TaskSuggestionsDismissParams,
+  TaskSuggestionsDismissResult,
+  TaskSuggestionsListParams,
+  TaskSuggestionsListResult,
+  TaskSummary,
+  TasksListParams,
+  TasksListResult,
+  TasksGetParams,
+  TasksGetResult,
+  TasksCancelParams,
+  TasksCancelResult,
+  CronJob,
+  CronListParams,
+  CronStatusParams,
+  CronGetParams,
+  CronAddParams,
+  CronAddResult,
+  CronDeclarativeAddResult,
+  CronUpdateParams,
+  CronRemoveParams,
+  CronRunParams,
+  CronRunsParams,
+  CronScratchGetParams,
+  CronScratchGetResult,
+  CronScratchSetParams,
+  CronScratchSetResult,
+  CronRunLogEntry,
+  ApprovalKind,
+  ApprovalDecision,
+  ApprovalAllowDecision,
+  ApprovalTerminalReason,
+  PluginApprovalSeverity,
+  ExecApprovalPresentation,
+  PluginApprovalPresentation,
+  ApprovalPresentation,
+  PendingApprovalSnapshot,
+  AllowedApprovalSnapshot,
+  DeniedApprovalSnapshot,
+  ExpiredApprovalSnapshot,
+  CancelledApprovalSnapshot,
+  ApprovalSnapshot,
+  TerminalApprovalSnapshot,
+  ApprovalGetParams,
+  ApprovalGetResult,
+  ApprovalHistoryParams,
+  ApprovalHistoryResult,
+  ApprovalResolveParams,
+  ApprovalResolveResult,
+  SessionApprovalEvent,
+  SessionApprovalReplay,
+  ExecApprovalsGetParams,
+  ExecApprovalsNodeSnapshot,
+  ExecApprovalsSetParams,
+  ExecApprovalsSnapshot,
+  ExecApprovalGetParams,
+  ExecApprovalRequestParams,
+  ExecApprovalResolveParams,
+  Question,
+  QuestionAnswers,
+  QuestionGetParams,
+  QuestionGetResult,
+  QuestionListParams,
+  QuestionListResult,
+  QuestionOption,
+  QuestionRecord,
+  QuestionRequestedEvent,
+  QuestionRequestParams,
+  QuestionRequestQuestion,
+  QuestionRequestResult,
+  QuestionResolvedEvent,
+  QuestionResolveParams,
+  QuestionResolveResult,
+  QuestionStatus,
+  QuestionWaitAnswerParams,
+  QuestionWaitAnswerResult,
+  LogsTailParams,
+  LogsTailResult,
+  TerminalOpenParams,
+  TerminalOpenResult,
+  TerminalInputParams,
+  TerminalUploadParams,
+  TerminalUploadResult,
+  TerminalResizeParams,
+  TerminalCloseParams,
+  TerminalAttachParams,
+  TerminalAttachResult,
+  TerminalSessionInfo,
+  TerminalListResult,
+  TerminalTextParams,
+  TerminalTextResult,
+  TerminalAckResult,
+  TerminalDataEvent,
+  TerminalExitEvent,
+  TerminalEvent,
+  PollParams,
+  WebPushVapidPublicKeyParams,
+  WebPushSubscribeParams,
+  WebPushUnsubscribeParams,
+  WebPushTestParams,
+  UpdateStatusParams,
+  UpdateRunParams,
+  ChatInjectParams,
+  WorktreeRecord,
+  WorktreesListParams,
+  WorktreesListResult,
+  WorktreesCreateParams,
+  WorktreesRemoveParams,
+  WorktreesRemoveResult,
+  WorktreesRestoreParams,
+  WorktreesGcParams,
+  WorktreesGcResult,
+  WorktreesBranchesParams,
+  WorktreeBranch,
+  WorktreeRepositoryStatus,
+  WorktreesBranchesResult,
+  FsDirEntry,
+  FsListDirParams,
+  FsListDirResult,
+  SessionGroup,
+  SessionsGroupsListParams,
+  SessionsGroupsListResult,
+  SessionsGroupsPutParams,
+  SessionsGroupsRenameParams,
+  SessionsGroupsDeleteParams,
+  SessionsGroupsMutationResult,
+} from "./schema.js";
+
+// Local structural result keeps this package independent of core session types.
+export type SessionsPatchResult = {
+  ok: true;
+  path: string;
+  key: string;
+  entry: Record<string, unknown>;
+  resolved?: {
+    modelProvider?: string;
+    model?: string;
+    agentRuntime?: GatewayAgentRuntime;
+    thinkingLevel?: string;
+    thinkingLevels?: Array<{ id: string; label: string }>;
+  };
+};
+
+type GatewayAgentRuntime = {
+  id: string;
+  fallback?: "openclaw" | "none";
+  source:
+    | "env"
+    | "agent"
+    | "defaults"
+    | "model"
+    | "provider"
+    | "implicit"
+    | "session"
+    | "session-key";
+};
